Extracting base client class WIP
